+++ packages/hasher/src/__tests__/resolveDependenciesHelper.ts
